Add additional animations (#102)
* Update OnDeckTIle.mat

* Add Character Idle Animation

Revised Story part 2 to match new ship design

* Add Character Idle Animation

Revised Story part 2 to match new ship design

* Set Default Character Material

Defaults to the main character's skin material. The other material can be applied when the prefab model is set into the scene, the dragging and dropping the new material onto the model portion of the prefab.

* Set Default Character Material

Defaults to the main character's skin material. The other material can be applied when the prefab model is set into the scene, the dragging and dropping the new material onto the model portion of the prefab.

* Update OnDeckTIle.mat

* Update OnDeckTIle.mat

* Add additional Animations

* Add additional Animations

* Add additional Animations

* Add Animation Events

Player character needs to be named "PlayerCharacter" for script to work.

Opponent character will occasionally play the thinking animation (even though it's not supposed to), but since it transitions better than the player character's version, that seems to be ok.

* Fix Animation Positioning

* Fix Animation Errors

* Minor fixes

* Update Story_Comic_part2.png